Move all supernodes to same package
