+++ src/trufflesom/compiler/ParserAst.java
@@ -44,12 +44,12 @@ import trufflesom.interpreter.nodes.literals.DoubleLiteralNode;
 import trufflesom.interpreter.nodes.literals.GenericLiteralNode;
 import trufflesom.interpreter.nodes.literals.IntegerLiteralNode;
 import trufflesom.interpreter.nodes.literals.LiteralNode;
-import trufflesom.interpreter.nodes.supernodes.LocalFieldStringEqualsNode;
-import trufflesom.interpreter.nodes.supernodes.LocalVariableSquareNodeGen;
-import trufflesom.interpreter.nodes.supernodes.NonLocalFieldStringEqualsNode;
-import trufflesom.interpreter.nodes.supernodes.NonLocalVariableSquareNodeGen;
-import trufflesom.interpreter.nodes.supernodes.StringEqualsNodeGen;
 import trufflesom.interpreter.supernodes.IntIncrementNodeGen;
+import trufflesom.interpreter.supernodes.LocalFieldStringEqualsNode;
+import trufflesom.interpreter.supernodes.LocalVariableSquareNodeGen;
+import trufflesom.interpreter.supernodes.NonLocalFieldStringEqualsNode;
+import trufflesom.interpreter.supernodes.NonLocalVariableSquareNodeGen;
+import trufflesom.interpreter.supernodes.StringEqualsNodeGen;
 import trufflesom.primitives.Primitives;
 import trufflesom.vm.Globals;
 import trufflesom.vm.NotYetImplementedException;
