+++ tests/trufflesom/supernodes/StringEqualsTests.java
@@ -7,8 +7,8 @@ import org.junit.Test;
 
 import trufflesom.interpreter.nodes.ExpressionNode;
 import trufflesom.interpreter.nodes.SequenceNode;
-import trufflesom.interpreter.nodes.supernodes.LocalFieldStringEqualsNode;
-import trufflesom.interpreter.nodes.supernodes.StringEqualsNode;
+import trufflesom.interpreter.supernodes.LocalFieldStringEqualsNode;
+import trufflesom.interpreter.supernodes.StringEqualsNode;
 import trufflesom.tests.AstTestSetup;
